+++ src/blockchain/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Export the Plugin
+ * @Author: JohnTrump
+ * @Date: 2019-08-19 15:52:46
+ * @Last Modified by: JohnTrump
+ * @Last Modified time: 2019-08-19 16:22:49
+ */
+import { EOS } from './eos'
+import { Cosmos } from './cosmos'
+import BlockchainPlugin from './Blockchain'
+import { SupportBlockchainEnums } from './SupportBlockchain'
+
+export { BlockchainPlugin, SupportBlockchainEnums }
+export { EOS, Cosmos }
