+++ src/util/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Export the utils
+ * @Author: JohnTrump
+ * @Date: 2019-08-19 15:52:46
+ * @Last Modified by: JohnTrump
+ * @Last Modified time: 2019-08-19 15:59:56
+ */
+
+import Network from './Network'
+import Tool from './Tool'
+
+export { Network, Tool }
